add: EitherSideSlider로 이름 변경 및 Slider가 들어간 차량 금액 뷰 생성
